% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{adjustLi}
\alias{adjustLi}
\title{Li correction of p-values}
\usage{
adjustLi(pvalues)
}
\arguments{
\item{pvalues}{Raw p-values in a matrix}
}
\value{
A matrix with the corrected p-values
}
\description{
This function takes the particular list of possible hypthesis to correct for multiple testing, as defined in Li, J. (2008)
}
\details{
The test has been implemented according to the version in Garcia et al. (2010), page 2680-2682.
}
\examples{
data(data.garcia.herrera)
raw.pvalues <- friedmanPost(data.garcia.herrera)
adjustLi(raw.pvalues)
}
\references{
Garcia S., Fernandez, A., Luengo, J. and Herrera, F. (2010) Advanced nonparametric tests for multiple comparison in the design of experiments in computational intelligence and data mining: Experimental analysis of power. \emph{Information Sciences}, 180, 2044-2064.
}

