% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scma.R
\name{all.vs.best.test}
\alias{all.vs.best.test}
\title{All vs. best statistical comparisons}
\usage{
\method{all}{vs.best.test}(results.matrix, test = wilcoxon.signed.test,
  group.by, alg.col, best = "max", summary = mean, correction = "holm",
  ...)
}
\arguments{
\item{results.matrix}{A matrix or data.frame containing the results obtained by the algorithms (columns) in each problem (rows).}

\item{test}{Statistical test to be applied. The first two arguments should be the two vectors to compare. The result has to be a list with an element named p.value. An example of this output is the typical \code{\link{htest}} class returned by R's statistical tests.}

\item{group.by}{A vector of names or indices of columns that will be used to group the results in \code{results.matrix}.}

\item{alg.col}{A vector of names or indices of columns indicating which columns contain the results for the algorithms to compare.}

\item{best}{A string indicating which should be considered the best result. Valid options are \code{'min'} and \code{'max'}.}

\item{summary}{A function used to summarize the data when looking for the best algorithm. By default, the median is used.}

\item{correction}{Type of correction to be applied to the p-values. Any method accepted by the \code{\link{p.adjust}} function is a valid option.}

\item{...}{Additional parameters to be passed to the test or the summarization function.}
}
\value{
A list with three matrices, \code{summary}, \code{raw.pvalues} and \code{adj.pvalues}. The first one contains the summarized values, the second one the raw p-values obtained in the comparison and the third one the adjusted pvalues. In the p-values matrices \code{NA} indicates the reference used in that row (i.e., the algorithms with the best value)
}
\description{
Given a results matrix, this function tests the differences between all the algorithms with respect to the best.
}
\examples{
dir <- system.file("loading_tests",package="scma")
file <- paste(dir , "beta_complete_comparison.out" , sep="/")
data <- read.comparison.file (file = file , alg.names = c('kakizawa','vitale','boundarykernel','betakernel'))
all.vs.best.test (data , group.by = c('size' , 'alpha' , 'beta') , alg.col=4:7 , test = t.test , best='min' , summary = mean , correction = 'hommel' , na.rm = TRUE , paired = TRUE)
}

