% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{anova.post}
\alias{anova.post}
\title{Tukey post hoc test for ANOVA}
\usage{
\method{anova}{post}(data, ...)
}
\arguments{
\item{data}{Data set (matrix or data.frame) to apply the test. The column names are taken as the groups and the values in the matrix are the samples}
}
\value{
A matrix with all the pairwise corrected p-values.
}
\description{
This function computes all the pairwise p-values corrected using Tukey post hoc test
}
\details{
The test has been implemented according to Test 22 in Kanji (2006).
}
\references{
Kanji, G. K. (2006) \emph{100 Statistical Tests}. SAGE Publications Ltd, 3rd edition.
}

