% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{anova.test}
\alias{anova.test}
\title{ANOVA test for multiple comparisons}
\usage{
\method{anova}{test}(data, ...)
}
\arguments{
\item{data}{Matrix where the test is performed}

\item{...}{Ignored}
}
\value{
A list with class "htest" containing the following components: \code{statistic}, the value of the statistic used in the test; \code{p.value}, the p-value for the test; \code{method}, a character string indicating what type of test was performed and \code{data.name}, a character string giving the name of the data.
}
\description{
This function performs F-test for K populations means
}
\details{
The test has been implemented according to Test 22 in Kanji (2006).
}
\examples{
data(data.garcia.herrera)
anova.test(data.garcia.herrera)
}
\references{
Kanji, G. K. (2006) \emph{100 Statistical Tests}. SAGE Publications Ltd, 3rd edition.
}

