% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{bCorrelatedTtest}
\alias{bCorrelatedTtest}
\title{Bayesian equivalent to the correlated t-test}
\usage{
bCorrelatedTtest(x, y = NULL, rho = 0, rope = c(-0.01, 0.01))
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample (if not provided, x is assumed to be the difference)}

\item{rho}{Correlation factor (see details)}

\item{rope}{Interval for the difference considered as "irrelevant"}
}
\value{
A list with the following elements: 
\item{\code{method}}{a string with the name of the method used}
\item{\code{posterior.probabilities}}{a vector with the left, rope and right probabilities}
\item{\code{approximated}}{a logical value, \code{TRUE} if the posterior distribution is approximated (sampled) and \code{FALSE} if it is exact}
\item{\code{parameters}}{parameters used by the method}
\item{\code{posterior}}{posterior density function if the method is exact and the obtained sample if the method is approximated}
\item{\code{additional}}{a list that contains the posterior cumulative function (\code{pposterior}), the posterior quantile function (\code{qposterior}) and the parameters of the posterior density function (\code{posterior.df}, \code{posterior.mean}, \code{posterior.sd})}
}
\description{
Implementation of the Bayesian version of the correlated t-test as presented in Benavoli \emph{et al.} 2017
}
\details{
Note that the default value for rho is 0, wich converts the test in the equivalent
of the standard t-test. To correct due to correlation you need to set the rho
parameter. In the case of classifiers compared using any validation scheme the
heuristic typically used is to set rho to num. test instances / total num. of instance
The function has been implemented to be used in the comparison of classifiers
and, in such situation, the heuristic used to fix the correlation factor
is the size of the training set divided by the total size of the data.
For the same reason, the reference value (the midpoint for the rope) is 0.
However, this may be changed (though it matches the prior for the paremter,
which follows a Gaussian distribution of 0 mean).
The results includes the typicall information relative to the three areas
of the posterior density (left, right and rope probabilities), but also
the basic functions (density, cummulative and quantile), as well as the
parameters of the posterior density function, which is a Student's t.
}
\examples{
sample1 <- rnorm(25, 1, 1)
sample2 <- rnorm(25, 1.2, 1)
correlatedTtest (x=sample1, y=sample2, rho=0.1, alternative="less")

}
\references{
A. Benavoli, G. Corani, J. Demsar, M. Zaffalon (2017) Time for a Change: a Tutorial
for Comparing Multiple Classifiers Through Bayesian Analysis. \emph{Journal of Machine Learning Research}, 18, 1-36.
}
