% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{bergmann.hommel.dynamic}
\alias{bergmann.hommel.dynamic}
\title{Bergmann and Hommel dynamic correction of p-values}
\usage{
bergmann.hommel.dynamic(raw.matrix)
}
\arguments{
\item{raw.matrix}{Raw p-values in a matrix}
}
\value{
A matrix with the corrected p-values
}
\description{
This function takes the particular list of possible hypthesis to correct for multiple testing, as defined in Bergmann and Hommel (1994)
}
\details{
The test has been implemented according to the version in Garcia and Herrera (2008), page 2680-2682.
}
\examples{
data(data.garcia.herrera)
raw.pvalues <- friedman.post(data.garcia.herrera)
bergmann.hommel.dynamic (raw.pvalues)
}
\references{
Garcia S. and Herrera, F. (2008) An Extension on "Statistical Comparisons of Classifiers over Multiple Data Sets" for All Pairwise Comparisons. \emph{Journal of Machine Learning Research}, 9, 2677-2694.
}

