% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{bHierarchicalTest}
\alias{bHierarchicalTest}
\title{Bayesian hierarchical model for the analysis of two algorithms in multiple datasets}
\usage{
bHierarchicalTest(x.matrix, y.matrix = NULL, rho, std.upper = 1000,
  d0.lower = NULL, d0.upper = NULL, alpha.lower = 0.5, alpha.upper = 5,
  beta.lower = 0.05, beta.upper = 0.15, rope = c(-0.01, 0.01),
  nsim = 2000, nchains = 8, parallel = TRUE, stan.output.file = NULL,
  seed = as.numeric(Sys.time()), ...)
}
\arguments{
\item{x.matrix}{First sample, a matrix with the results obtained by the first algorithm (each dataset in a row)}

\item{y.matrix}{Second sample, a matrix with the results obtained by the second algorithm (each dataset in a row) (if not provided, x is assumed to be the difference)}

\item{std.upper}{Factor to set the upper bound for both sigma_i and sigma_0 (see Benavoli \emph{et al.} 2017 for more details)}

\item{d0.lower}{Lower bound for the prior for mu_0. If not provided, the smallest observed difference is used}

\item{d0.upper}{Upper bound for the prior for mu_0. If not provided, the biggest observed difference is used}

\item{alpha.lower}{Lower bound for the (uniform) prior for the alpha hyperparameter (see Benavoli \emph{et al.} 2017 for more details). Default value set at 0.5, as in the original paper}

\item{alpha.upper}{Upper bound for the (uniform) prior for the alpha hyperparameter (see Benavoli \emph{et al.} 2017 for more details). Default value set at 5, as in the original paper}

\item{beta.lower}{Lower bound for the (uniform) prior for the beta hyperparameter (see Benavoli \emph{et al.} 2017 for more details). Default value set at 0.05, as in the original paper}

\item{rope}{Interval for the difference considered as "irrelevant"}

\item{nsim}{Number of samples (per chain) used to estimate the posterior distribution. Note that, by default, half the simulations are used for the burn-in}

\item{parallel}{Logical value. If \code{true}, Stan code is executed in parallel}

\item{stan.output.file}{String containing the base name for the output files produced by Stan. If \code{NULL}, no files are stored.}

\item{seed}{Optional parameter used to fix the random seed}

\item{...}{Additional arguments for the rstan::stan function that runs the analysis}

\item{beta.lower}{Upper bound for the (uniform) prior for the beta hyperparameter (see Benavoli \emph{et al.} 2017 for more details). Default value set at 0.15, as in the original paper}

\item{z0}{Position of the pseudo-observation associated to the prior Dirichlet Process. The default value is set to 0 (inside the rope)}

\item{nchain}{Number of MC chains to be simulated. As half the simulations are used for the warm-up, the total number of simulations will be \code{nchain}*\code{nsim}/2}
}
\value{
A list with the following elements: 
\item{\code{method}}{a string with the name of the method used}
\item{\code{parameters}}{parameters used by the method}
\item{\code{posterior.probabilities}}{a vector with the left, rope and right probabilities}
\item{\code{approximated}}{a logical value, \code{TRUE} if the posterior distribution is approximated (sampled) and \code{FALSE} if it is exact}
\item{\code{posterior}}{Sampled probabilities (see details)}
\item{\code{additional}}{Additional information provided by the model. This includes:\code{per.dataset}, the results per dataset (left, rope and right probabilities together with the expected mean value); \code{global.sin} sampled probabilities of mu_0 being positive or negative and \code{stan.results}, the complete set of results produced by Stan program}
}
\description{
Bayesian hierarchical model for the simulatenous analysis of two algorithms in multiple datasets as presented in Benavoli \emph{et al.} 2017
}
\details{
The results includes the typical information relative to the three 
areas of the posterior density (left, right and rope probabilities), both global and per dataset (in the additional information). Also, the simulation results are included.

As for the prior parameters, they are set to the default values indicated in Benavoli \emph{et al.} 2017, except for the bound for the prior distribution of mu_0, which are set to the maximum and minimum values observed in the sample. You should not modify them unless you know what you are doing.
}
\examples{
sample1 <- matrix(rnorm(25*5, 1, 1), nrow=5)
sample2 <- matrix(rnorm(25*5, 1.2, 1), nrow=5)
results <- bHierarchicalTest (x.matrix=sample1, y.matrix=sample2, rho=0, rope=c(-0.05, 0.05))
res$posterior.probabilities

}
\references{
A. Benavoli, G. Corani, J. Demsar, M. Zaffalon (2017) Time for a Change: a Tutorial
for Comparing Multiple Classifiers Through Bayesian Analysis. \emph{Journal of Machine Learning Research}, 18, 1-36.
}
