% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{bPlackettLuceModel}
\alias{bPlackettLuceModel}
\title{Bayesian Plackett-Luce model for ranking analysis}
\usage{
bPlackettLuceModel(x.matrix, min = TRUE, prior = rep(1, ncol(x.matrix)),
  nsim = 2000, nchains = 8, parallel = TRUE, stan.output.file = NULL,
  seed = as.numeric(Sys.time()), ...)
}
\arguments{
\item{x.matrix}{Sample of performance of the algorithms. Each column is an algorithm and each row is a problem}

\item{min}{Logical value indicating which values should be ranked in first postion. If \code{TRUE}, the smallest value will be the first ranked and if \code{FALSE} the first will be the one with the highest value}

\item{prior}{Hyperparameters of the prior distribution of the weights. It should be a vector of size equal to the number of algorithms of real valued numbers greater than 0By default, all equal to 1}

\item{parallel}{Logical value. If \code{true}, Stan code is executed in parallel}

\item{stan.output.file}{String containing the base name for the output files produced by Stan. If \code{NULL}, no files are stored.}

\item{seed}{Optional parameter used to fix the random seed}

\item{...}{Additional arguments for the rstan::stan function that runs the analysis}

\item{nchain}{Number of MC chains to be simulated. As half the simulations are used for the warm-up, the total number of simulations will be \code{nchain}*\code{nsim}/2}
}
\value{
A list with the following elements: 
\item{\code{method}}{a string with the name of the method used}
\item{\code{parameters}}{parameters used by the method}
\item{\code{posterior.weights}}{a vector with the weights sampled from the posterior distribution}
\item{\code{expected.win.prob}}{for each algorithm, the expected posterior probability of being the best}
\item{\code{expceted.mode.rank}}{for each algorithm the expected rank in the most probable ranking}
\item{\code{additional}}{complete results produced by the Stan program}
}
\description{
Bayesian model based on the Plackett-Luce distribution over rankings to analyse multiple algorithms in multiple problems
}
\examples{
n.alg  <- 5
n.inst <- 25
x.matrix <- matrix(runif(n.alg*n.inst), ncol=n.alg)
colnames(x.matrix) <- paste("Alg", 1:n.alg, sep="")
rownames(x.matrix) <- paste("Inst", 1:n.inst, sep="")

res <- bPlackettLuceModel(x.matrix, min=FALSE, nsim=2000, nchains=3)

res$expected.win.prob
res$expected.mode.rank
}
