% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scmamp.R
\name{contrastEstimationMatrix}
\alias{contrastEstimationMatrix}
\title{Contrast estimation based on medians}
\usage{
contrastEstimationMatrix(data)
}
\arguments{
\item{data}{Matrix or data frame with the data to compare}
}
\value{
A matrix where the estimation of all the pairs of differences are output.
The differences correspond to row-column.
}
\description{
This function performs estimates the contrast between algorithms through the medians
}
\details{
The test has been implemented according to Garcia \emph{et al.} (2010), Section 3.3.
}
\examples{
data(data_gh_2008)
contrastEstimationMatrix(data.gh.2008)
}
\references{
Kanji, G. K. (2006) \emph{100 Statistical Tests}. SAGE Publications Ltd, 3rd edition.
}

