% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{correct.for.monotocity}
\alias{correct.for.monotocity}
\title{Enforce the monotocity of a sequence}
\usage{
correct.for.monotocity(pvalues)
}
\arguments{
\item{pvalues}{Vector where the correction has to be performed}
}
\value{
A vector with the corrected values
}
\description{
This function ensures that the i-th element is not smaller than any previous one
}
\details{
In some corrections it can happen that after the correction the order of the resulting p-values is not the same as in the raw pvalues. A very simple example is the Holm correction where the ordered p-values are multiplied by m, m-1, ... , 2, 1. In case the last two p-values are, for instance, 0.7 and 0.8, the corrected p-values would be ... , 0.7*2, 0.8*1, i.e., ... , 1 , 0.8. This situation needs a correction to set the last value at 1.
}
\examples{
pv <- c(4.4870e-07,1.0416e-06,0.00116,0.0288,0.032, 0.0303 , 0.0384 , 0.0384 , 1 , 1)
correct.for.monotocity(pv)
}

