% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{correlatedTtest}
\alias{correlatedTtest}
\title{Correlated t-test}
\usage{
correlatedTtest(x, y = NULL, rho, alternative = "two.sided")
}
\arguments{
\item{x}{First sample}

\item{y}{Second sample (if not provided, x is tested aginst 0)}

\item{rho}{Correlation factor (see details)}

\item{alternative}{a character string specifying the alternative hypothesis, 
must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}
}
\value{
A list with class "htest" containing the following components: \code{statistic}, the value of the statistic used in the test; \code{method}, a character string indicating what type of test was performed; \code{data.name}, a character string giving the name of the data and \code{diff.matirx}, a matrix with all the pairwise absolute difference of average values.
}
\description{
This function performs the correlated version of the t-test for 
paired samples
}
\details{
This version of the t-test takes into account the correlation between samples.
for the particular case of cross validated results, the heuristic to set the correlation is
size of test set divided by total size of the dataset.
}
\examples{
sample1 <- rnorm(25, 1, 1)
sample2 <- rnorm(25, 1.2, 1)
correlatedTtest (x=sample1, y=sample2, rho=0.1, alternative="less")
}
\references{
C. Nadeau and Y. Bengio (2003) Inference for the Generalization Error. \emph{Machine Learning}, 52(3), 239-281.
}
