% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{critical.difference.plot}
\alias{critical.difference.plot}
\title{Critical difference plot}
\usage{
critical.difference.plot(results.matrix, alpha = 0.05, cex = 0.75)
}
\arguments{
\item{results.matrix}{Matrix or data frame with the results for each algorithm}

\item{alpha}{Significance level to get the critical difference. By default this value is 0.05}

\item{cex}{Numeric value to control the size of the font. By default it is set at 0.75.}
}
\description{
This function plots the critical difference plots shown in \emph{Demsar (2006)}
}
\examples{
data(data.garcia.herrera)
pwcomp.bh <- pairwise.test(results.matrix = data.garcia.herrera , test = "Friedman post" , correction = "Bergmann Hommel")
critical.difference.plot(data.garcia.herrera)
}
\references{
Demsar, J. (2006) Statistical Comparisons of Classifiers over Multiple Data Sets. \emph{Journal of Machine Learning Research}, 7, 1-30.
}
\seealso{
\code{\link{algorithm.graph}}, \code{\link{plot.pvalues}}
}

