% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{customPost}
\alias{customPost}
\title{Function to use custom tests to perform pairwise comparisons}
\usage{
customPost(data, control, test, ...)
}
\arguments{
\item{data}{Data set (matrix or data.frame) to apply the test. The column names are taken as the groups and the values in the matrix are the samples}

\item{test}{Function to perform the test. It requires two parameters, \code{x} and \code{y}, the two samples to be compared, and it has to return a list that contains, at least, one element called p.value (as the \code{htest} objects that are usually returned by R's statistical test implementations)}

\item{...}{Additional parameters for the test function.}
}
\value{
A matrix with all the pairwise raw p-values.
}
\description{
This function computes the raw p-values for all the pairwise comparisons using a custom function
}
\examples{
data(data.garcia.herrera)
test <- function(x, y, ...) {
  t.test(x, y, paired=TRUE)
}
customPost(data.garcia.herrera , test)
}

