% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scmamp.R
\name{data.blum.2015}
\alias{data.blum.2015}
\title{Comparison of optimization algorithms in Blum \emph{et al.} (2015)}
\format{A data frame with 10 columns and 900 rows}
\source{
C. Blum, B. Calvo and M.J. Blesa (2015)  FrogCOL and FrogMIS: New Decentralized Algorithms for Finding Large Independent Sets in Graphs. \emph{Swarm Intelligence}. In press.
}
\description{
This dataset contains part of the results obtained in the comparison of decentralyzed
optimization algorithms presented in Blum \emph{et al.} (2015). The dataset contains
900 rows and 10 colums. Each row reprsents an instance of the maximum independent
set problem (a graph). The first two are descriptors of the problem in each
row (size and radius used to create random geometric graphs) and the other 8
contain the results obtained by 8 algorithms for the MIS problem instance.
}

