% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scmamp.R
\name{data.garcia.herrera}
\alias{data.garcia.herrera}
\title{Example in Garcia and Herrera (2008)}
\format{A data frame with 5 columns and 30 rows}
\source{
García S. and Herrera, F. (2008) An Extension on "Statistical Comparisons of Classifiers over Multiple Data Sets" for all Pairwise Comparisons. \emph{Journal of Machine Learning Research}. 9, 2677-2694.
}
\description{
Dataset corresponding to the accuracy of 5 classifiers in 30 datasets
Each algorithm is in a column
}

