% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scmamp.R
\name{data.gh.2008}
\alias{data.gh.2008}
\title{Example in Garcia and Herrera (2008)}
\format{A data frame with 5 columns and 30 rows}
\source{
S. Garcia and F. Herrera (2008) An Extension on "Statistical
Comparisons of Classifiers over Multiple Data Sets" for all Pairwise
Comparisons. \emph{Journal of Machine Learning Research}. 9, 2677-2694.
}
\description{
Dataset corresponding to the accuracy of 5 classifiers in 30 datasets.
Each algorithm is in a column. This is the dataset used as example in
Garcia and Herrera (2008).
}

