% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scmamp.R
\name{data.gh.2010}
\alias{data.gh.2010}
\title{Example in Garcia and Herrera (2010)}
\format{A data frame with 4 columns and 24 rows}
\source{
S. Garcia and F. Herrera (2010) Advanced Nonparametric Tests for 
Multiple Comparison in the Design of Experiments in Computational Intelligence 
and Data Mining: Experimental Analysis of Power. \emph{Information Sciences}, 
180, 2044-2064.
}
\description{
Dataset corresponding to the accuracy of 4 classifiers in 24 datasets.
Each algorithm is in a column. This is the dataset used as example in 
Garcia and Herrera (2010).
}

