% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pvalue_corrections.R
\name{exhaustive.sets}
\alias{exhaustive.sets}
\title{Create the complet set of exhaustive sets}
\usage{
exhaustive.sets(set)
}
\arguments{
\item{set}{Set to create the exhaustive sets}
}
\value{
A list with all the possible exhaustive sets, without repetitions
}
\description{
This function implements the algorithm in Figure 1, Garcia and Herrera (2008) to create, given a set, the complete set of exhaustive sets E
}

