% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{exhaustiveSets}
\alias{exhaustiveSets}
\title{Create the complet set of exhaustive sets}
\usage{
exhaustiveSets(set)
}
\arguments{
\item{set}{Set to create the exhaustive sets. The complexity of this algorithm is huge, so use with caution for sets of more than 7-8 elements.}
}
\value{
A list with all the possible exhaustive sets, without repetitions
}
\description{
This function implements the algorithm in Figure 1, Garcia and Herrera (2008) to create, given a set, the complete set of exhaustive sets E
}
\examples{
exhaustiveSets(c("A","B","C","D"))
}

