% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{friedman.post}
\alias{friedman.post}
\title{Friedman's post hoc raw p-values}
\usage{
friedman.post(data, ...)
}
\arguments{
\item{data}{Data set (matrix or data.frame) to apply the test. The column names are taken as the groups and the values in the matrix are the samples}
}
\value{
A matrix with all the pairwise raw p-values.
}
\description{
This function computes the raw p-values for the post hoc based on Friedman's test
}
\details{
The test has been implemented according to the version in Demsar (2006), page 12
}
\references{
Demsar, J. (2006) Statistical Comparisons of Classifiers over Multiple Data Sets. \emph{Journal of Machine Learning Research}, 7, 1-30.
}

