% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{friedman.test}
\alias{friedman.test}
\title{Friedman's test}
\usage{
friedman.test(data, ...)
}
\arguments{
\item{data}{Matrix where the test is performed}

\item{...}{Ignored}
}
\value{
A list with class "htest" containing the following components: \code{statistic}, the value of the statistic used in the test; \code{parameter}, the two degrees of freedom of the F distribution; \code{p.value}, the p-value for the test; \code{method}, a character string indicating what type of test was performed and \code{data.name}, a character string giving the name of the data.
}
\description{
This function performs Friedman's test
}
\examples{
data("garcia.herrera")
iman.davenport.test(data.garcia.herrera)
}

