% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{friedmanAlignedRanksPost}
\alias{friedmanAlignedRanksPost}
\title{Friedman's aligned ranks post hoc raw p-values for all vs. control and all vs. all comparisons}
\usage{
friedmanAlignedRanksPost(data, control = NULL, ...)
}
\arguments{
\item{data}{Data set (matrix or data.frame) to apply the test. The column names are taken as the groups and the values in the matrix are the samples.}

\item{control}{Either the number or the name of the column for the control algorithm. If this parameter is not provided, the all vs all comparison is performed.}

\item{...}{Not used.}
}
\value{
A matrix with all the pairwise raw p-values (all vs. all or all vs. control).
}
\description{
This function computes the raw p-values for the post hoc based on Friedman's test.
}
\details{
The test has been implemented according to the version in Garcia et al. (2010), pages 2051,2054
}
\examples{
data(data.garcia.herrera)
friedmanAlignedRanksPost(data.garcia.herrera)
}
\references{
Garcia, S. et al. (2010) Advanced nonparametric tests for multiple comparisons in the design of experiments in computational intelligence and ata mining: Experimental analysis of power. \emph{Information Sciences}, 180, 2044-2064.
}

