% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tests.R
\name{nemenyi.cd}
\alias{nemenyi.cd}
\title{Nemenyi test critical difference}
\usage{
nemenyi.cd(alpha = 0.05, num.alg, num.problems)
}
\arguments{
\item{alpha}{Significance of the test}

\item{num.alg}{Number of algorithms (treatments, etc.)}

\item{num.problems}{Number of problems (samples)}
}
\value{
Critical difference of averge rankings. When the difference is, in absolute value, greater than this value the null hypothesis (no differences) is rejected.
}
\description{
This function computes the critical difference for the Nemenyi test
}

