% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scma.R
\name{pairwise.test}
\alias{pairwise.test}
\title{Statistical pairwise comparisons of algorithms}
\usage{
pairwise.test(results.matrix, test = "Wilcoxon", correction = "Shaffer",
  ...)
}
\arguments{
\item{results.matrix}{A matrix or data.frame containing the results obtained by the algorithms (columns) in each problem (rows).}

\item{test}{Parameter that indicates the statistical test to be used. It can be either a string indicating one of the available test (\code{t-test} and \code{Wilcoxon}) or a function that, given two parameters, \code{x} and \code{y}, return the p-value associated with the comparison. For compatibility, this function has to have the ... special argument.}

\item{correction}{A string indicating the type of correction that has to be applied to the p-values. Valid names are \code{Shaffer}, \code{Bergmann-Hommel}, \code{Nemenyi}, \code{Tukey} or any of the methods implemented in the \code{p.adjust} function. For a list of options, type \code{p.adjust.methods}}

\item{...}{Special argument used to pass additional parameters to the statistical test or the correction method.}
}
\value{
The function returns a list that contains:
\itemize{
\item {\code{raw.pvalues} - Matrix with the raw p-values}
\item {\code{corrected.pvalues} - Matrix with the corrected p-values}
\item {\code{test} - String character indicating the test used}
\item {\code{correction} - String character indicating the correction used}
}
}
\description{
This function gives access to different alternatives to test the algorithms pair-wise
}
\details{
The most powerfull method is the dynamic procedure by Bergmann and Hommel, but its computational requirements render this method only applicable for few algorithms. In the current version of the package it accepts up to 8 classifiers. Shaffer's static approach increases the power without much less cost. Regarding the Nemenyi test, it is equivalent to using Bonferroni's correction implemented in the \code{'mt.raw2adjp'} function. Note that Tukey post hoc test is designed for ANOVA and, thus, if selected the \code{test} parameter is ignores. Simalrly, the Nemenyi test is the non parametric version and thus, it also ignores the \code{test} parameter.
}
\seealso{
\code{plot.pvalues}, \code{plot.hypothesis}, \code{algorithm.graph} , \code{scma}
}

