% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{partition}
\alias{partition}
\title{Complete set of (non-empty) partitions of a set}
\usage{
partition(set)
}
\arguments{
\item{set}{Set to be partitioned}
}
\value{
All the possible partitions of the set where the last element is in the second subset and the first subset is not empty.
}
\description{
This function creates all the possible partitions indicated in the 6th step of the algorithm shown in Figure 1, Garcia and Herrera (2008).
}

