% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.densities}
\alias{plot.densities}
\title{Kernel based density estimation of the samples}
\usage{
\method{plot}{densities}(results.matrix, ...)
}
\arguments{
\item{results.matrix}{A matrix where columns represent the algorithms}

\item{...}{The plot is created using \code{\link{ggplot2}}. This special parameter can be used to pass additional parameters to the \code{\link{geom_line}} function used to plot the sample points.}
}
\value{
A \code{\link{ggplot}} object.
}
\description{
This function estimates and plots the densities of the results of each algorithm
}
\examples{
data(data.garcia.herrera)
plot.densities(data.garcia.herrera)
}
\seealso{
\code{\link{gaussian.qqplot}}
}

