% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.pvalues}
\alias{plot.pvalues}
\title{Plotting the p-value matrix}
\usage{
\method{plot}{pvalues}(pvalue.matrix, alg.order = NULL, show.pvalue = TRUE,
  font.size = 5)
}
\arguments{
\item{pvalue.matrix}{Matrix with the p-values to plot}

\item{alg.order}{A permutation indicating the reordering of the algorithms}

\item{show.pvalue}{Logical value indicating whether the numerical values have to be printed}

\item{font.size}{Size of the p-values, if printed}
}
\value{
A \code{\link{ggplot}} object.
}
\description{
This function plots the p-value matrix as a \code{\link{ggplot2}}'s tile plot.
}
\examples{
data(data.garcia.herrera)
pwcomp.bh <- pairwise.test(results.matrix = data.garcia.herrera , test = "Friedman post" , correction = "Bergmann Hommel")
plot.pvalues(pwcomp.bh$corrected.pvalues)
}
\seealso{
\code{\link{algorithm.graph}}, \code{\link{critical.difference.plot}}
}

