% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBarycentric}
\alias{plotBarycentric}
\title{Plotting barycentric plots}
\usage{
plotBarycentric(data.matrix)
}
\arguments{
\item{data.matrix}{a matrix or dataframe where each row is an n-dimension data point (whose sum is 1)}
}
\value{
An object of class \linkS4class{ggplot} with the plot
}
\description{
This function plots tuples of vectors that sum 1 in barycentric coordinates
}
\examples{
data <- matrix(runif(100*5), ncol=5)
data <- data / rowSums(data)
plotBarycentric(data)

}
