% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDensities}
\alias{plotDensities}
\title{Kernel based density estimation of the samples}
\usage{
plotDensities(data, ...)
}
\arguments{
\item{data}{A matrix where columns represent the algorithms}

\item{...}{The plot is created using \code{ggplot2}. This special parameter can be used to pass additional parameters to the \code{\link{geom_line}} function used to plot the sample points. It can also be used to pass additional arguments to the \code{density} function, which is used to eastimate the densities.}
}
\value{
A \code{\link{ggplot}} object.
}
\description{
This function estimates and plots the densities of the results of each algorithm
}
\examples{
data(data_gh_2010)
plotDensities(data.gh.2010)

}
\seealso{
\code{\link{qqplotGaussian}}
}

