% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPosterior}
\alias{plotPosterior}
\title{Plotting the (marginal) posterior densities in Bayesian analyses}
\usage{
plotPosterior(results, parameter = 1, num.points = 1000, plot.rope = TRUE,
  plot.samples = TRUE, alpha = NULL, ...)
}
\arguments{
\item{results}{A list containing, at least three elements, one named \code{approximate}, which is
a logical value indicating whether the posterior is a function or a sample, \code{rope}, a two dimensional
vector with the minimum and maximum values of the rope and \code{posterior}, either a one parameter function or
a matrix (or data.frame) where each row is a sample and each column a sampled parameter}

\item{parameter}{Either a string or a number indicating, in case the posterior is approximated, the parameter
to be ploted (i.e., the name or the index of a column in the sample matrix)}

\item{num.points}{Number of points used to plot the functions}

\item{plot.rope}{A logical value indicating whether the rope has to be plotted or not. Note that not for all
parameter the rope makes sense}

\item{plot.samples}{A logical value. If true, the samples are plotted (only when the posterior is approximate)}

\item{alpha}{Numeric value for the transparency of the points, only applicable if \code{plot.samples} is true}

\item{...}{Additional parameters to the Rgraphviz function. This is mainly to change the layout of the graph}
}
\value{
An object of class \linkS4class{ggplot} with the plot
}
\description{
This function plots, univariately, the posterior densities obtained
in Bayesian analyses.
}
\details{
Note that if the methods are exact (not simulated), the true density can be plotted but,
for those cases where the posterior is approximated through sampling, the function will
plot a kernel density estimation of the posterior and, thus, the probabilities computed
by other functions are not directly the areas under the densities.
}
\examples{
x <- rnorm(25, 1, 2)
y <- rnorm(25, 1.1, 2)
results <- bCorrelatedTtest(x=x, y=y, rho=0, rope=c(-0.05, 0.05))
plotPosterior(results)

}
