% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{read.comparison.dir}
\alias{read.comparison.dir}
\title{Read data from a directory of comparison-like files}
\usage{
read.comparison.dir(directory, alg.names, col.names, names, fname.pattern, ...)
}
\arguments{
\item{directory}{Directory where the files are located.}

\item{alg.names}{List of the column names that include the results. The rest are assumed as descriptors of the problems.}

\item{col.names}{Vector of names of the columns. If not NULL, the files are assumed not to have a header and the columns are named using this vector.}

\item{names}{List of names for the variables to be extracted from the file name.}

\item{fname.pattern}{Regular expression to extract information from the file names. It has to be a regular expression that matches the name of the files and where the information to be extrcted has to be between brakets. As an example, if the whole file name wants to be used, the expression \code{'([.]*)'} can be used. For more example see the examples below or the vignette covering the data loading.}

\item{...}{Additional parameters for the read.csv function used to load the data. It can be used, for example, to set the separator (e.g., \code{sep="\t"}).}
}
\value{
A data.frame where each column represents either a feature of the experiment or the result of running an algorithm. Algorithm columns are placed always at the end of the table.
}
\description{
This function reads the data from all files in a directory. Each file is expected to to be formated as a comparison file, i.e., the file can have some columns that characterize the problem and one column per algorithm. If each row contain only the result obtained by one algorithm, use the \code{\link{read.experiment.dir}} function.
}
\examples{
dir <- paste(system.file("loading_tests",package="scma") , "comparison_files" , sep="/")
## The format of the files is beta_ALPHA,BETA_size_SIZE.out, where variables to extract are in capital letters.
list.files(dir)[1]
## The regular expresion can ba as simple as substituting each variable name in the expression above by ([XXX]*), where XXX is the list of symbols that appear in the name.
pattern <- "beta_([0-9]*),([0-9]*)_size_([0-9]*).out"
var.names <- c('alpha' , 'beta' , 'size')
alg.names <- c('kakizawa','vitale','boundarykernel','betakernel')
data <- read.comparison.dir (directory = dir , alg.names = alg.names , col.names = NULL , names = names , fname.pattern = pattern)
dim(data)
head(data)
}
\seealso{
\code{\link{read.experiment.file}}, \code{\link{read.experiment.dir}}, \code{\link{read.comparison.dir}}
}

