% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loading_functions.R
\name{read.experiment.dir}
\alias{read.experiment.dir}
\title{Read data from an experiment-like files in a directory}
\usage{
read.experiment.dir(directory, names, alg.var.name, value.col, fname.pattern,
  ...)
}
\arguments{
\item{directory}{Directory with the files to load. It should only contain files to load, no other kind of file.}

\item{names}{List of names for the variables to be extracted from the file name}

\item{alg.var.name}{Name of the variable that defines the algorithm used in the experiment. It can be either one of the variables extracted from the file name or the header of the data in the file.}

\item{value.col}{Name or index (referred to the column in the file) of the column indicating the value for the algorithm}

\item{fname.pattern}{Regular expression to extract information from the file names. It has to be a regular expression that matches the name of the files and where the information to be extrcted has to be between brakets. As an example, if the whole file name wants to be used, the expression \code{'([.]*)'} can be used. For more example see the examples below or the vignette covering the data loading.}

\item{...}{Additional parameters for the read.csv function used to load the data. It can be used, for example, to set the separator (e.g., \code{sep="\t"}).}
}
\value{
A data.frame where each column represents either a feature of the experiment or the result of running an algorithm. Algorithm columns are placed always at the end of the table.
}
\description{
This function reads the data from all the files in a directory. Only one column is expected in each file, correspondig to the results obtained by a single algorithm. If the files contain the results of two or more algorithms, see function \code{\link{read.comparison.file}}. The function can extract information from the file name.
}
\details{
Note that all the files should have the same format (only one column with the same header)
}
\examples{
dir <- paste(system.file("loading_tests",package="scmamp") , "experiment_files" , sep="/")
## The format of the files is beta_ALPHA,BETA_size_SIZE_ESTIMATOR.out, where variables to extract are in capital letters.
list.files(dir)[1]
## The regular expresion can ba as simple as substituting each variable name in the expression above by ([XXX]*), where XXX is the list of symbols that appear in the name.
pattern <- "beta_([0-9]*),([0-9]*)_size_([0-9]*)_([a-z]*).out"
var.names <- c('alpha' , 'beta' , 'size', 'estimator')
data <- read.experiment.dir (dir , var.names , 'estimator' , pattern)
dim(data)
head(data)
}
\seealso{
\code{\link{read.experiment.file}}, \code{\link{read.comparison.file}}, \code{\link{read.comparison.dir}}
}

