% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{recursive.count}
\alias{recursive.count}
\title{Maximum number of true hypothesis}
\usage{
recursive.count(k)
}
\arguments{
\item{k}{Number of algorithms}
}
\value{
List of maximum number of true hypothesis in a pair-wise comparsion of \code{k} classifiers
}
\description{
This function gets the S(k) set as described in Shaffer (1985)
}
\examples{
recursive.count(5)
}

