% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scma.R
\docType{package}
\name{scma}
\alias{scma}
\alias{scma-package}
\title{Statistical comparison of multiple algorithms}
\description{
This package has been develop to simplify the statistical assessment of algorithms when tested in different problems. It includes statistical tests, as well as some plotting functions.
}
\author{
Borja Calvo, Guzman Santafe
}

