% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scmamp.R
\docType{package}
\name{scmamp}
\alias{scmamp}
\alias{scmamp-package}
\title{Statistical comparison of multiple algorithms}
\description{
This package has been develop to simplify the statistical assessment of algorithms when tested in different problems. It includes statistical tests, as well as some plotting functions.
}
\author{
Borja Calvo \email{borja.calvo@ehu.eus},
Guzman Santafe \email{guzman.santafe@unavarra.es}

Maintainer: Borja Calvo \email{borja.calvo@ehu.es}
}
\seealso{
For an overview of the use see #' \code{vignette(topic=
"Statistical_comparison_of_multiple_algorithms_in_multiple_problems",
package="scmamp")} and \code{vignette(topic="Data_loading_and_manipulation",
package="scmamp")}
}

