% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pvalue_corrections.R
\name{shaffer.static}
\alias{shaffer.static}
\title{Shaffer's correction of p-values in pair-wise comparisons}
\usage{
shaffer.static(raw.matrix)
}
\arguments{
\item{raw.matrix}{A matrix with the pair-wise p-values. The p-values have to be, at least, in the upper part of the matrix.}
}
\value{
A symetric matrix with the corrected p-values.
}
\description{
This function implements the Shaffer's multiple testing correction when the p-values correspond with pair-wise comparisons
}

