% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pvalue_corrections.R
\name{subdivisions}
\alias{subdivisions}
\title{All the ordered subdivisions of a set}
\usage{
subdivisions(set)
}
\arguments{
\item{set}{Set to be subdivided}
}
\value{
All the possible subdbisions of the set, including those where there are empty sets, but without repetitions. The format is a list of lists. Each element in the main list is a list containing two vectors, \code{s1} and \code{s2}, the two subsets of the set passed in the argument
}
\description{
This function is the base to crate all the partitions needed in the algorithm to create all the exhaustive sets in Figure 1, Garcia and Herrera (2008).
}

