% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{summarize.data}
\alias{summarize.data}
\title{Summarization of data}
\usage{
summarize.data(data, fun = mean, group.by = NULL, ignore = NULL, ...)
}
\arguments{
\item{data}{A matrix or data frame to be filtered.}

\item{fun}{Function to be used in the summarization. It can be any function that taking as first argument a numeric vector otuputs a numeric value. Typical examples are \code{\link{mean}}, \code{\link{median}}, \code{\link{min}}, \code{\link{max}} or \code{\link{sd}}.}

\item{group.by}{A vector of either column names or column indices according to which the data will be grouped to be summarized.}

\item{ignore}{A vector of either column names or column indices that will be ignored (not printed out).}

\item{...}{Additional parameters to the summarization function (\code{fun}). For example, \code{na.rm = TRUE} to indicate that the missing values should be ignored.}
}
\value{
A data frame where, for each combination of the values in the columns indicated by \code{group.by}, each column (except those in \code{ignore}) contains the summarization of the values in the original matrix that have that combination of values.
}
\description{
This is a simple function to apply a summarization function to a matrix or data frame.
}
\examples{
dir <- system.file("loading_tests",package="scmamp")
file <- paste(dir , "beta_complete_comparison.out" , sep="/")
data <- read.comparison.file (file = file , alg.names = c('kakizawa','vitale','boundarykernel','betakernel'))
summarize.data (data , group.by = c('size','alpha','beta') , ignore = c(4,5,7) , fun = median)
## group by size. Get the mean and variance
summarize.data (data , group.by = 1 , ignore = 2:3 , fun = mean , na.rm = TRUE)
summarize.data (data , group.by = 1 , ignore = 2:3 , fun = sd , na.rm = TRUE)
}
\seealso{
\code{\link{filter.data}}
}

