% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pvalue_corrections.R
\name{tukey.test}
\alias{tukey.test}
\title{Tukey post hoc test for ANOVA}
\usage{
tukey.test(data, ...)
}
\arguments{
\item{data}{Data set (matrix or data.frame) to apply the test. The column names are taken as the groups and the values in the matrix are the samples}
}
\value{
A matrix with all the pairwise corrected p-values.
}
\description{
This function computes all the pairwise p-values corrected using Tukey post hoc test
}

