% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pvalue_corrections.R
\name{unique.exhaustive.sets}
\alias{unique.exhaustive.sets}
\title{Remove repetitions in a list of subsets}
\usage{
\method{unique}{exhaustive.sets}(E)
}
\arguments{
\item{E}{list of exhaustive sets}
}
\value{
The list without repetitions
}
\description{
This function removes the repetitions in the result returned by the function \code{\link{exhaustive.sets}}
}

