% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/latex.R
\name{write.tabular}
\alias{write.tabular}
\title{Write a table in LaTeX format}
\usage{
write.tabular(table, file = NULL, format = "g", bold = NULL,
  italic = NULL, mark = NULL, mark.char = "*", align = "l",
  hrule = NULL, vrule = NULL, bty = c("t", "b", "l", "r"),
  print.col.names = TRUE, print.row.names = TRUE, digits = rep(3,
  dim(table)[2] + print.row.names))
}
\arguments{
\item{table}{A data frame with the information to plot}

\item{file}{Path of a file. If provided, the tabular is wirten in the given file. Otherwise, it is writen to the standard output}

\item{format}{Format for the numeric values. The accepted formats are those in the function \code{\link{formatC}}. The typical values are \code{'g'} to automatically set theformat, \code{'f'} for a fixed sized floating point format and \code{'e'} or \code{'E'} for scientific notation}

\item{bold}{A matrix that matches \code{'table'} in size indicating with true those cells that have to be printed in bold font}

\item{italic}{A matrix that matches \code{'table'} in size indicating with true those cells that have to be printed in bold font}

\item{mark}{A matrix that matches \code{'table'} in size indicating with true those cells that have to be marked with a superscipt symbol}

\item{mark.char}{Character to be used to mark cells. Note that the superscript is included in a math environment, so this has to be either a character or a valid math command in LaTeX}

\item{align}{Character indicating the alignment of the colums (\code{'l'},\code{'r'} or \code{'c'})}

\item{hrule}{A vector of positions for the horizontal lines in the tabular. All the lines are drawn after the indicated line. When the column names are plotted, 0 means drawing a line after the column names. The maximum value is the number of rows - 1 (for a line after the last line see parametr \code{bty})}

\item{vrule}{Similar to \code{'hrule'} but for vertical lines. . The maximum value is the number of columns - 1 (for a line after the last columns see parametr \code{bty})}

\item{print.col.names}{Local value indicating whether the column names have to be printed or not}

\item{print.row.names}{Local value indicating whether the row names have to be printed or not}

\item{digits}{A vector with the number of digits in each column. Its size has to match the number of the final table, i.e., the colums in \code{'table'} if the row names are not included or the number of columns + 1 if the row names are printed in the final table}
}
\value{
LaTeX code to print the table
}
\description{
This is a simple function to create tabular environment in LaTeX
}
\examples{
data(data.garcia.herrera)
## Fixed width, 4 digits
ncol <- dim(data.garcia.herrera)[2]
digits <- rep(4,ncol+1)
write.tabular(data.garcia.herrera , format = 'f' , hrule = 0 , vrule = 0 , digits = digits)
## Scientific notation, maximum value in each row in bold
max.matrix <- t(apply(data.garcia.herrera , MARGIN = 1 , FUN = function(x) x==max(x)))
write.tabular(data.garcia.herrera , format = 'E' , bold = max.matrix)
}

