% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pvalue_corrections.R
\name{bergmann.hommel.dynamic}
\alias{bergmann.hommel.dynamic}
\title{Bergmann and Hommel dynamic correction of p-values}
\usage{
bergmann.hommel.dynamic(raw.matrix)
}
\arguments{
\item{raw.matrix}{Raw p-values in a matrix}
}
\value{
A matrix with the corrected p-values
}
\description{
This function takes the particular list of possible hypthesis to correct for multiple testing, as defined in Bergmann and Hommel (1994)
}

