% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scma.R
\name{pairwise.test}
\alias{pairwise.test}
\title{Statistical pairwise comparisons of algorithms}
\usage{
pairwise.test(results.matrix, test = "Friedman post",
  correction = "Shaffer", ...)
}
\arguments{
\item{results.matrix}{A matrix or data.frame containing the results obtained by the algorithms (columns) in each problem (rows).}

\item{test}{Parameter that indicates the statistical test to be used. It can be either a string indicating one of the available test (\code{'t-test'} for paired t-test,  \code{'Wilcoxon'}) for Wilcoxon Signed rank test, \code{'Friedman post'}, for raw p-values in Friedman test's Nemenyi post hoc or a test or a function with, at least, two parameters, \code{x} and \code{y}, which are the two samples to be compared. The function has to return a list that contains, at least, one element called p.value (as the \code{htest} objects that are usually returned by R's test implementations).}

\item{correction}{A string indicating the type of correction that has to be applied to the p-values. Valid names are \code{Shaffer}, \code{Bergmann-Hommel}, \code{Nemenyi}, \code{Tukey} or any of the methods implemented in the \code{p.adjust} function. For a list of options, type \code{p.adjust.methods}.}

\item{...}{Special argument used to pass additional parameters to the statistical test or the correction method.}
}
\value{
The function returns a list that contains:
\itemize{
\item {\code{raw.pvalues} - Matrix with the raw p-values}
\item {\code{corrected.pvalues} - Matrix with the corrected p-values}
\item {\code{test} - String character indicating the test used}
\item {\code{correction} - String character indicating the correction used}
}
}
\description{
This function gives access to different alternatives to test the algorithms pair-wise
}
\details{
The most powerfull method is the dynamic procedure by Bergmann and Hommel, but its computational requirements render this method only applicable for few algorithms. In the current version of the package it accepts up to 9 algorithms. If selected in a dataset with more than 9 columns the correction is automatically changed to \code{'Shaffer'} and a warning is displayed. Shaffer's static approach increases the power without much less cost. Regarding the Nemenyi test, when used with \code{'Friedman post'} test it is equivalent to using Bonferroni's correction implemented in the \code{'mt.adj'} function. Note that Tukey post hoc test is designed for ANOVA and, thus, if selected the \code{test} parameter has to be \code{'t-test'}; otherwise a warning is shown and the \code{test} parameter is changed. Simalrly, the Nemenyi test has to be coupled with Friedman test's post, and if not it is modified an a warning is displayed.
}
\examples{
data(data.garcia.herrera)
pwcomp.bh <- pairwise.test(results.matrix = data.garcia.herrera , test = "Friedman post" , correction = "Bergmann Hommel")
plot.pvalues(pwcomp.bh$corrected.pvalues)
}
\seealso{
\code{plot.pvalues}, \code{plot.hypothesis}, \code{algorithm.graph}.
}

