% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post_hoc.R
\name{shaffer.static}
\alias{shaffer.static}
\title{Shaffer's correction of p-values in pair-wise comparisons}
\usage{
shaffer.static(raw.matrix)
}
\arguments{
\item{raw.matrix}{A matrix with the pair-wise p-values. The p-values have to be, at least, in the upper part of the matrix.}
}
\value{
A symetric matrix with the corrected p-values.
}
\description{
This function implements the Shaffer's multiple testing correction when the p-values correspond with pair-wise comparisons
}
\details{
The test has been implemented according to the version in Garcia and Herrera (2008), page 2680.
}
\examples{
data(data.garcia.herrera)
raw.pvalues <- friedman.post(data.garcia.herrera)
shaffer.static (raw.pvalues)
}
\references{
Garcia S. and Herrera, F. (2008) An Extension on "Statistical Comparisons of Classifiers over Multiple Data Sets" for All Pairwise Comparisons. \emph{Journal of Machine Learning Research}, 9, 2677-2694.
}

